Move colorFromString to formHelpers
